feat: Hosts create and update operations dialog
