+++ app/dashboard/src/pages/hosts/dialog/default-values.ts
@@ -1,0 +1,14 @@
+import { HostSchema } from 'stores';
+export const getDefaultValues = (): HostSchema => {
+    return {
+        host: '',
+        sni: '',
+        port: null,
+        path: '',
+        address: '',
+        remark: '',
+        security: 'inbound_default',
+        alpn: '',
+        fingerprint: '',
+    }
+};
